--- client/App.js
@@ -1,13 +1,0 @@
-import "bootstrap/dist/css/bootstrap.min.css";
-
-function App() {
-  return (
-    <div>
-      <div>This is a message board</div>
-      <div>This is a user list</div>
-      <div>This is a message box</div>
-    </div>
-  );
-}
-
-export default App;
