--- client/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-  <head>
-    <title>Smalltalk</title>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
-  </head>
-  
-  <body>
-    <div id="root"></div>
-  </body>
-  
-</html>
