+++ frontend/public/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+
+  <head>
+    <meta charset="utf-8" />
+    <title>React App</title>
+  </head>
+  
+  <body>
+    <div id="root"></div>
+  </body>
+  
+</html>
